+++ src/Core/SharedKernels/Enumerator/Enumerator/ViewModels/InterviewDetails/Questions/MultiOptionQuestionViewModel.cs
@@ -85,7 +85,7 @@ namespace WB.Core.SharedKernels.Enumerator.ViewModels.InterviewDetails.Questions
             this.Options = new ReadOnlyCollection<MultiOptionQuestionOptionViewModel>(
                 questionnaire
                 .GetAnswerOptionsAsValues(entityIdentity.Id)
-                .Select(x => new CategoricalOption { Value = Convert.ToInt64(x), Title = questionnaire.GetAnswerOptionTitle(entityIdentity.Id, x) })
+                .Select(x => new CategoricalOption { Value = Convert.ToInt32(x), Title = questionnaire.GetAnswerOptionTitle(entityIdentity.Id, x) })
                 .Select((x, index) => this.ToViewModel(x, existingAnswer, index))
                 .ToList());
         }
