+++ src/Core/SharedKernels/Enumerator/Enumerator/ViewModels/InterviewDetails/Questions/SingleOptionQuestionViewModel.cs
@@ -79,7 +79,7 @@ namespace WB.Core.SharedKernels.Enumerator.ViewModels.InterviewDetails.Questions
 
             this.Options = questionnaire
                 .GetAnswerOptionsAsValues(entityIdentity.Id)
-                .Select(x => new CategoricalOption { Value = Convert.ToInt64(x), Title = questionnaire.GetAnswerOptionTitle(entityIdentity.Id, x) })
+                .Select(x => new CategoricalOption { Value = Convert.ToInt32(x), Title = questionnaire.GetAnswerOptionTitle(entityIdentity.Id, x) })
                 .Select(model => this.ToViewModel(model, isSelected: model.Value == selectedValue))
                 .ToList();
             this.eventRegistry.Subscribe(this, interviewId);
